# A faster version of exp, exp2, exp10 for Floating point numbers