… base/special/exp10.jl