@oscardssmith 2020-07-22
Doctests seem to be failing due to a floating point precision issue. Any recommendations?