@jakobnissen 2020-07-22
You can change the docstring. I'm nearly 100% certain that Julia explicitly does not guarantee specific floating point rounding is preserved across minor versions.