@chriselrod 2020-07-22
Motivation to get something like it, `gexp` is my translation of the GLIBC implementation (that I wrote by looking at the GPL source, meaning its definition must be GPL-ed):
```julia
julia> using LoopVectorization

julia> x = rand(256); y = similar(x);

julia> @btime @. $y = exp($x); # base
  1.458 μs (0 allocations: 0 bytes)

julia> @btime @avx @. $y = exp($x);  # ccall
  141.684 ns (0 allocations: 0 bytes)

julia> @btime @avx @. $y = gexp($x); # Julia translation
  83.055 ns (0 allocations: 0 bytes)
```
I didn't implement the range check, so an actual implementation would probably be a bit slower.