@kimikage 2020-07-22
BTW, in general, it is not fair to use random numbers in `[0, 1)` for benchmarking `exp`. Some algorithms rely heavily on the branch prediction.