@oscardssmith 2020-07-22
Do you have a recommendation on how to get a good test set?