@kimikage 2020-07-22
I have no particular recommendations, but I think `randn` is simple and slightly fairer than `rand`.