@chriselrod 2020-07-22
```julia
julia> @btime @. $y = exp($x);
  1.696 μs (0 allocations: 0 bytes)

julia> @btime @avx @. $y = exp($x);
  180.336 ns (0 allocations: 0 bytes)

julia> @btime @avx @. $y = gexp($x);
  81.509 ns (0 allocations: 0 bytes)

julia> @btime @. $y = myexp($x);
  608.153 ns (0 allocations: 0 bytes)
```
`exp` is the current version, and `myexp` is from this PR. It is failing to SIMD, but the scalar performance is impressive.
I (or someone else) can do an accuracy comparison later.