@musm 2020-07-22
I'm against table lookups because they are hard to use with SIMD, both Yepp [1]  and SLEEF avoid them due to this important reason. 

[1] https://pdfs.semanticscholar.org/f993/776749e3a3e12836a1802b985cd7b524653d.pdf


> Four design principles for portable high-throughput vector elementary
> functions motivated by processor microarchitecture.
> I No table lookups
> I No division or square root operations
> I No branches except for special cases
> I Use Horner Scheme and Newton-Raphson with Software Pipelining


Importantly regarding accuracy:
Have you tested this is  <= 1 ULP  over the FULL domain.


 GPL source:  Can we even use this if you looked at the source (I'm not an expert on this)