@chriselrod 2020-07-22
> I'm against table lookups because they are hard to use with SIMD, both Yepp [1] and SLEEF avoid them due to this important reason.

Both `@avx` versions I benchmarked above are SIMD and use table lookups.