@musm 2020-07-22
Can you post your `gexp` (or put it in a package)?

A table lookup is probably always going to be faster for a single value. I think the current version will still triumph in overall speed and accuracy considering both single elements and vectorized elements. Note I spent **an absurd amount of time** exploring various different implementations and versions and the on in in base Julia is in my opinion a very **very very** solid implementation.