@chriselrod 2020-07-22
> Can you post your gexp (or put it in a package or gist...probably the best idea)?

I'll make a package for these in a couple weeks, but here is a gist for now:
https://gist.github.com/chriselrod/d5cbc1dc7f2427cea86080d33a0cf5dc

> I think the current version will still triumph in overall speed and accuracy considering both single elements and vectorized elements.

For some reason performance of `gexp` seems to have regress from 80->100 ns when switching from LLVM10 to LLVM9 for 256 doubles on my computer, but it is far ahead in terms of speed when vectorized:
```julia
julia> using LoopVectorization, BenchmarkTools, Test

julia> x = 50 .* randn(256); y = similar(x);

julia> @btime @. $y = exp($x); # current version
  1.468 μs (0 allocations: 0 bytes)

julia> @btime @avx @. $y = exp($x); # GLIBC
  147.725 ns (0 allocations: 0 bytes)

julia> @btime @avx @. $y = gexp($x); # GLIBC translated to Julia, but leaving off range checks
  100.337 ns (0 allocations: 0 bytes)

julia> @test y == (@avx exp.(x))
Test Passed

julia> @btime @. $y = myexp($x); # This PR
  608.676 ns (0 allocations: 0 bytes)
```

In your experience, how much performance tends to be lost to gain a few more ULP in accuracy?