@oscardssmith 2020-07-22
From my testing, it doesn't appear that this PR introduces extra ULPs of error.