@musm 2020-07-22
```julia

julia> @btime exp($(Ref(2.0))[])
  9.109 ns (0 allocations: 0 bytes)


julia> @btime gexp($(Ref(2.0))[])
  311.245 ns (5 allocations: 80 bytes)

julia>  x = 10000 .* randn(256); y = similar(x);


julia>  @btime @. $y = exp($x);
  1.080 μs (0 allocations: 0 bytes)

julia>  @btime @. $y = gexp($x);
  12.699 μs (63 allocations: 1008 bytes)


```