@oscardssmith 2020-07-22
For measuring this PR, you want `exp` not `gexp` (that is the vectorized version that I haven't yet finished). I get
```
julia> @btime myexp($(Ref(2.0))[]);
  2.234 ns (0 allocations: 0 bytes)

julia> @btime exp($(Ref(2.0))[]);
  6.141 ns (0 allocations: 0 bytes)
```