@oscardssmith 2020-07-22
I have absolutely no idea how this would be taking 300ns for scalar or allocating. If you look at the code, there's nothing in there capable of causing an allocation.