@musm 2020-07-22
Looks like the table look up was missing `Float64` that was corrected by a comment that I missed.