@chriselrod 2020-07-31
Making the function `@inline` so that it can SIMD and copy pasting the required definitions for running the code:
```julia
# -log(2)/512 into upper and lower bits
LN2o512U(::Type{Float64}) = -0.0013538030870311431
LN2o512L(::Type{Float64}) = -4.529388308293554e-20
# 256/log(2)
LN2o512INV(::Type{Float64}) = 738.6598609351493

# magic rounding constant: 1.5*2^52 Adding, then subtracting it from a float rounds it to an Int.
MAGIC_ROUND_CONST(::Type{Float64}) = 6755399441055744.0

# max and min arguments
MAX_EXP(::Type{Float64}) = 709.782712893383996732   # log 2^1023*(2-2^-52)
# one less than the min exponent since we can sqeeze a bit more from the exp function
MIN_EXP(::Type{Float64}) = -745.1332191019412076235   # log 2^-1075

@inline expm1_kernel(x) = x * evalpoly(x,  (0.9999999999999994,
											0.4999999999999998,
											0.1666666714395386,
											0.04166666778033678))

const J_TABLE = Float64[2.0^(-53 + big(j-1)/512) for j in 1:512]

@inline function myexp(x::T) where T<:Float64
    if !(abs(x) < MAX_EXP(T))
        x >= MAX_EXP(T) && return Inf
        x <= MIN_EXP(T) && return 0.
        isnan(x) && return x
    end
    
    N_float = muladd(x, LN2o512INV(T), MAGIC_ROUND_CONST(T))
    N = reinterpret(Int64, N_float) % Int32
    N_float -=  MAGIC_ROUND_CONST(T)
    rhi = muladd(N_float, LN2o512L(T), muladd(N_float, LN2o512U(T), x))
    k = N >> 9
    jt_val = @inbounds J_TABLE[N&511 + 1]
    small_part =  muladd(jt_val, expm1_kernel(rhi), jt_val)
    twopk = rem(53 + k, UInt64) << 52
    return reinterpret(T, twopk+reinterpret(Int, small_part))
end
```

In terms of runtime performance:
```julia
julia> x = rand(256); y = similar(x);

julia> @benchmark myexp!($y, $x)
 BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     204.937 ns (0.00% GC)
  median time:      207.452 ns (0.00% GC)
  mean time:        207.891 ns (0.00% GC)
  maximum time:     427.346 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     555

julia> function simdivdepf!(f, y, x)
           @inbounds @simd ivdep for i ∈ eachindex(y,x)
               y[i] = f(x[i])
           end
       end
 simdivdepf! (generic function with 1 method)

julia> @benchmark simdivdepf!(myexp, $y, $x)
 BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     207.357 ns (0.00% GC)
  median time:      208.198 ns (0.00% GC)
  mean time:        208.411 ns (0.00% GC)
  maximum time:     277.445 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     560

julia> @benchmark simdivdepf!(exp, $y, $x)
 BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.214 μs (0.00% GC)
  median time:      1.223 μs (0.00% GC)
  mean time:        1.225 μs (0.00% GC)
  maximum time:     2.222 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     10
```
This is about 6 times faster on my computer.
Without the `@simd ivdep`, it is closer to 2.3x faster:
```julia
julia> @benchmark $y .= myexp.($x)
 BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     530.679 ns (0.00% GC)
  median time:      534.458 ns (0.00% GC)
  mean time:        535.076 ns (0.00% GC)
  maximum time:     842.516 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     190
```

Testing accuracy with:
```julia
julia> xx = range(log(floatmin(Float64)), log(floatmax(Float64)), length = 10^7);

julia> fun_table = Dict(myexp => Base.exp)
 Dict{typeof(myexp),typeof(exp)} with 1 entry:
  myexp => exp

julia> tol = 10
 10

julia> @time test_acc(Float64, fun_table, xx, tol, debug = true, tol_debug = 3)
Tol debug failed 0.0% of the time.
myexp             : max 0.999800 at x = 515.57044                       : mean 0.336230
Test Summary:  | Pass  Total
accuracy myexp |    1      1
 29.599016 seconds (143.91 M allocations: 6.548 GiB, 1.07% gc time)
 1-element Vector{Any}:
 Test.DefaultTestSet("accuracy myexp", Any[], 1, false)
```
Maximum of 0.9998.